Add status indicator
