+++ apps/ledger-live-mobile/src/components/AnalyticsConsole/Status.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Flex, Text } from "@ledgerhq/native-ui";
+import { useSelector } from "react-redux";
+import { getIsTracking } from "../../analytics";
+
+const trackingSelector = state => getIsTracking(state, false);
+
+const Status: React.FC<{}> = () => {
+  const isTracking = useSelector(trackingSelector);
+  return (
+    <Flex>
+      <Text color="black">
+        Tracking {isTracking.enabled ? "enabled" : "disabled"}
+        {!isTracking.enabled ? ": " + isTracking.reason : ""}
+      </Text>
+    </Flex>
+  );
+};
+
+export default Status;
